+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `arachni_rpcd` -- Fixed bug causing the `--nickname` option to not be understood.
 - `UI::Output` -- Flush output stream after each message.
+- `Check::Auditor#skip?` -- Ignore mutations when checking for redundancies.
 - `Browser` -- Fixed issue causing `select` inputs in forms to not be set.
 - `Element::Cookie.encode` -- Added '&' to the list of reserved characters.
 - `Issue`